파이어스토어 관련 로직 별도 함수로 추출
